requirements.txt
streamlit>=1.37.0
litellm>=1.43.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.3
tiktoken>=0.7.0
rapidfuzz>=3.9.0
